fix datatype
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@42164 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/smbd/smbd/buffer_pool.c
@@ -27,7 +27,7 @@ struct wm {
 
 struct wm_list {
 	struct list_head	list;
-	unsigned int		sz;
+	size_t			sz;
 
 	spinlock_t		wm_lock;
 	int			avail_wm;
